refactor: change models name in greetings folder
