+++ internal/greetings/greetings.service.go
@@ -1,4 +1,4 @@
-package greeting
+package greetings
 
 import (
 	"fmt"
@@ -7,8 +7,8 @@ import (
 )
 
 type Service interface {
-	Greeting(*models.Greeting)
-	UserGreeting(*models.Greeting, string)
+	Greeting(*models.Greetings)
+	UserGreeting(*models.Greetings, string)
 }
 
 type serviceImpl struct{}
@@ -17,10 +17,10 @@ func NewService() Service {
 	return &serviceImpl{}
 }
 
-func (s *serviceImpl) Greeting(msg *models.Greeting) {
+func (s *serviceImpl) Greeting(msg *models.Greetings) {
 	msg.Message = "Hello, World!"
 }
 
-func (s *serviceImpl) UserGreeting(msg *models.Greeting, email string) {
+func (s *serviceImpl) UserGreeting(msg *models.Greetings, email string) {
 	msg.Message = fmt.Sprintf("Hello, %v!", email)
 }
